// src/api/generated/models/UserResponse.ts
export interface UserResponse {
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    esntlId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    userName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    groupId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    groupName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    branchId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    branchName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    positionId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    positionName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    managerId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    managerName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    statusCode?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    useAt?: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {Array<RoleInfo>}
     * @memberof UserResponse
     */
    roles?: Array<RoleInfo>;
    /**
     * 
     * @type {Date}
     * @memberof UserResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof UserResponse
     */
    updatedDate?: Date;
}