// src/api/generated/models/PageUserResponse.ts
export interface PageUserResponse {
    /**
     * 
     * @type {number}
     * @memberof PageUserResponse
     */
    totalPages?: number;
    /**
     * 
     * @type {number}
     * @memberof PageUserResponse
     */
    totalElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageUserResponse
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {number}
     * @memberof PageUserResponse
     */
    numberOfElements?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageUserResponse
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageUserResponse
     */
    last?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PageUserResponse
     */
    size?: number;
    /**
     * 
     * @type {Array<UserResponse>}
     * @memberof PageUserResponse
     */
    content?: Array<UserResponse>;
    /**
     * 
     * @type {number}
     * @memberof PageUserResponse
     */
    number?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageUserResponse
     */
    sort?: SortObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageUserResponse
     */
    empty?: boolean;
}